+++ package.json
@@ -14,6 +14,8 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-hook-form": "^6.3.3",
+    "react-modal": "^3.11.2",
+    "react-modal-hook": "^3.0.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "react-toastify": "^6.0.8"
