+++ src/components/Input/styles.css
@@ -1,6 +1,6 @@
 .input {
   position: relative;
-  padding: 0.2rem 5rem;
+  padding: 0.2rem 2rem;
 }
 
 .input label {
