+++ src/components/Modal/styles.css
@@ -1,0 +1,67 @@
+.modal {
+  position: absolute;
+  background-color: #fff;
+}
+
+.modal__overlay {
+  z-index: 1;
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(0, 0, 0, 0.6);
+}
+
+.modal__overlay > div {
+  overflow-y: scroll;
+  max-height: 100%;
+  width: 100%;
+  top: 50%;
+  left: 50%;
+  right: auto;
+  bottom: auto;
+  margin-right: -50%;
+  transform: translate(-50%, -50%);
+}
+
+.modal__overlay > div:focus {
+  border: none;
+}
+
+.modal__close {
+  width: 100%;
+  padding: 1rem 1rem 0 0;
+  font-size: 2rem;
+  text-align: right;
+}
+
+.modal__title {
+  color: #f1a10a;
+  font-size: 2rem;
+  font-weight: bold;
+  text-align: center;
+  width: 100%;
+  padding-bottom: 1rem;
+}
+
+@media all and (min-width: 720px) {
+  .modal__overlay > div {
+    max-height: 80%;
+    width: 70%;
+  }
+}
+
+@media all and (min-width: 960px) {
+  .modal__overlay > div {
+    max-height: 80%;
+    width: 70%;
+  }
+}
+
+@media all and (min-width: 1500px) {
+  .modal__overlay > div {
+    max-height: 70%;
+    width: 40%;
+  }
+}
