+++ src/components/Textarea/styles.css
@@ -1,6 +1,6 @@
 .textarea {
   position: relative;
-  padding: 0.2rem 5rem;
+  padding: 0.2rem 2rem;
 }
 
 .textarea label {
