+++ src/containers/ContactForm/styles.css
@@ -1,27 +1,11 @@
-.contact {
-  max-width: 1024px;
-  margin: 140px auto;
-  background: #fff;
-}
-
 .contact__fieldset {
   border: 0;
   padding: 0 2.4rem;
 }
 
-.contact__fieldset h1 {
-  color: #f1a10a;
-  height: 5rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  padding-bottom: 1rem;
-}
-
 .contact__fieldset .input + .input,
 .contact__fieldset .input + .textarea {
-  margin-top: 1.4rem;
+  margin-top: 1rem;
 }
 
 .contact__submit {
