+++ src/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import 'normalize.css';
-import './styles.css';
 import App from './App';
 
 ReactDOM.render(
